requirements.txt
qrcode[pil]>=7.4.2
opencv-python>=4.8.0
pillow>=9.5.0
pandas>=2.0.0
click>=8.1.0
numpy>=1.24.0
matplotlib>=3.7.0
Flask
